Add modules and intra.js
