+++ src/index.html
@@ -25,6 +25,6 @@
             <button class="clearBtn" id="clearButton">Clear all completed</button>
         </div>
     </section>
-    <script src="./index.js"></script>
+    <script type="" src="./index.js"></script>
 </body>
 </html>
