+++ webpack.config.js
@@ -5,6 +5,7 @@ module.exports = {
   mode: 'development',
   entry: {
     index: './src/index.js',
+   
   },
   devServer: {
     static: './dist',
